Fix Integer Size

+++ database/migrations/2018_08_13_000000_create_assets_table.php
@@ -17,7 +17,7 @@ class CreateAssetsTable extends Migration
             $table->increments('id');
             $table->string('type');
             $table->string('name');
-            $table->unsignedInteger('issuance');
+            $table->unsignedBigInteger('issuance');
             $table->boolean('divisible');
             $table->timestamps();
         });
